Remove unused Minecraft bot and models
Deleted Minecraft bot script and two related model files. The bot script was removed as it is outdated and no longer being utilized in the current application. The model files are large, obsolete and not necessary for our current model training, thus they were also removed to clean up the codebase and reduce repository size.