Update Minecraft model and fine-tune its performance
This commit modifies the Minecraft model (model.pth). It has made necessary changes to improve the performance of the model. It has altered FloatStorage and metadata to fix some weight and bias initialization issues. The linear1 and linear2 layers have been adjusted to better accommodate the game's parameters. Moreover, it has improved version compatibility. The goal of these changes was to make the model more efficient and accurate during gameplay.