Refactor score calculation in Minecraft agent
The method of calculating the mean score in the Minecraft agent has been changed. Instead of dividing the total score by the number of games which might lead to incorrect results if games were skipped or not included in the score, it now divides the total score by the length of plot_scores which includes all the played games in it. This results in more accurate scoring data.