Add new Minecraft model file
This commit adds a new model file that is used for the Minecraft implementation. The model is a Pytorch model and is stored in a .pth format.

This change was required to provide a model that can be used to understand and predict the player's actions based on the given input data. The model reflects a series of transformations and computations that have been trained and